+++ RELEASE_NOTES.md
@@ -14,8 +14,8 @@
 ### Breaking Changes
 
 * `PhysicsEngine` should no longer be instantiated in the Server `main.js` and in the client entry point.  Rather, it should be instantiated in the `GameEngine`.
+* `PhysicsEngine` constructor now does initialization instead of the `init` function
 * `GameEngine` step method cannot be called without passing the `isReenact` argument.  Games which override the `step` method must pass this argument when calling the super method.
-
 
 ## Release 1.0.1